Add functions to convert hex and format input
`hexbin` returns converts a hex to bin and returns it as a string.
`getinstr` takes this string as input and splits it into MIPS
instruction segments. It then returns an array containing the format (R
or I), the opcode name, and the arguments.

Currently only works with R-format instructions (more specifically the
subset used for the project)